fix: flip signal direction to 'long' — engine emits only negative-funding signals
The live engine emits regime_updated events ONLY for negative-funding
assets (regime_detector.py:163-165 filters `if funding < 0`;
trading_engine.py:806-807 filters `if funding >= 0: continue`). On a
negative-funding asset, the earning side is LONG (longs collect when
funding is negative; shorts pay).

Before this commit:
- CompositeSignalScorer hardcoded direction="short" on every signal
- LiveOrchestrator hardcoded direction="short" when opening positions
- ScoredSignal default was "short", executor fallback was "short"

Net effect: every live paper/real trade went the WRONG way. Shorts on
negative-funding assets PAY funding, mirror-imaging the backtester's
82.6% win rate into an expected ~17.4%. This explains the growing
negative funding_usd observed on open YZY/DOT paper positions.

Changes:
- src/scoring/composite_scorer.py: emit direction="long"
- src/pipeline/live_orchestrator.py: read signal.direction, fall back to "long"
- src/models.py: ScoredSignal.direction default = "long"
- src/execution/executor.py: fallback default = "long"
- tests/test_direction_sign.py: new regression suite (8 tests) covering
  model default, scorer output, orchestrator wiring, and all four
  (direction × rate-sign) funding-accrual combinations
- tests/test_execution.py, tests/test_live_orchestrator.py: updated
  pre-existing assertions that locked in the wrong-direction behavior

When the upstream engine eventually propagates a signed funding rate
per asset, switch to the signed derivation directly (matching
scripts/backtest/strategies/funding_arb.py:72's
`"short" if rate > 0 else "long"`).

+++ src/execution/executor.py
@@ -379,14 +379,16 @@ class Executor:
         context["size_coins"] = size_coins
         context["leverage"] = self.leverage
 
-        # Resolve trade direction from the signal. Default to "short" if
-        # the signal didn't carry one — that matches the backtester /
-        # paper-trader convention for HIGH_FUNDING (positive funding,
-        # short earns). Inverting this on a real-money trade would
-        # cause us to PAY funding instead of collecting it.
-        direction = (getattr(signal, "direction", None) or "short").lower()
+        # Resolve trade direction from the signal. Default to "long" if
+        # the signal didn't carry one — the live engine only emits
+        # NEGATIVE-funding signals (regime_detector filters `if funding < 0`;
+        # trading_engine scanner filters `if funding >= 0: continue`). On a
+        # negative-funding asset the earning side is LONG (longs collect
+        # when funding is negative). Inverting this on a real-money trade
+        # would cause us to PAY funding instead of collecting it.
+        direction = (getattr(signal, "direction", None) or "long").lower()
         if direction not in ("long", "short"):
-            direction = "short"
+            direction = "long"
         is_buy = direction == "long"
         context["direction"] = direction
